Money attributes changed to BigDecimal, which is generally recommended

+++ src/main/java/at/ac/tuwien/ec/mongouk2011/entities/WorkerEntity.java
@@ -1,5 +1,6 @@
 package at.ac.tuwien.ec.mongouk2011.entities;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -10,11 +11,11 @@ public class WorkerEntity extends EmployeeEntity {
 
 	private Integer yearsExperience;
 	
-	public WorkerEntity() {
+	public WorkerEntity(){
 		super();
 	}
 	public WorkerEntity(String firstname, String surname, List<String> telephone, List<String> fax,
-			List<String> mobile, String email, Double salary, Integer yearsExperience) {
+			List<String> mobile, String email, BigDecimal salary, Integer yearsExperience) {
 		super(firstname, surname, telephone, fax, mobile, email, salary);
 		this.yearsExperience = yearsExperience;
 	}
